refactor: split builtins_multidim_ops into three modules
`src/runtime/builtins_multidim_ops.rs` had reached 978 lines, nearly twice
CLAUDE.md's 500-line-per-file convention and over it by a wide margin for a
long time. Split along the builtin-name groupings that already organised the
file, each new module lands comfortably under the limit:

- `builtins_multidim_ops.rs` (273) keeps what all three need --
  `resolve_multidim_indices`, `writeback_multidim_var_to_local`, the
  `multidim_empty_list`/`multidim_missing_result` miss-shape rules with their
  raku-vs-roast evidence, `check_shaped_index_bounds` -- plus
  `builtin_multidim_delete`, which is not an adverb family of its own and is
  the only caller of the bounds check.
- `builtins_multidim_subscript_adverb.rs` (360) takes the `:v`/`:k`/`:p`/`:kv`
  family and its two dynamic-`:$delete` twins.
- `builtins_multidim_exists_adverb.rs` (385) takes the `:exists` family and
  the two nested-slice support functions.

No logic changed. Every handler kept the exact `pub(super)` visibility that
`src/runtime/builtins.rs`'s dispatcher relies on. Two items widened because a
private `fn` in an `impl` block is private to its module and their callers now
live in sibling modules: `resolve_multidim_indices` and the two free functions
`multidim_empty_list`/`multidim_missing_result`, all private -> `pub(super)`.
Each family's `_multi` helper and `check_shaped_index_bounds` stayed private.

Verified with the targeted roast sweep the ticket asked for -- all 57
whitelisted S09-typed-arrays / S32-array / S32-hash / multidim / multislice
files (9375 tests), including both multislice-6e.t files that pin the
miss-shape rules -- plus the full local t/ suite (3495 files, 34516 tests).
