+++ src/runtime/builtins_multidim_ops.rs
@@ -1,15 +1,24 @@
+//! Shared support for the multidim (`;`-separated) subscript builtins, plus
+//! the `:delete` handler.
+//!
+//! The adverb handlers themselves live in the two sibling modules
+//! `builtins_multidim_subscript_adverb` (`:v`/`:k`/`:p`/`:kv` and their
+//! dynamic-`:$delete` twins) and `builtins_multidim_exists_adverb`
+//! (`:exists` and its secondary adverbs). Everything three of them share --
+//! index resolution, the miss-shape rules, the local-slot writeback -- is
+//! here.
+
 use super::*;
 use crate::value::ArrayKind;
 
 use super::builtins_multidim::{
-    array_to_list, has_multi_indices, leaf_key_tuple, make_key_tuple, multidim_collect_leaves,
-    multidim_delete, multidim_index, multidim_index_with_hole,
+    array_to_list, has_multi_indices, multidim_collect_leaves, multidim_delete, multidim_index,
 };
 
 /// The empty-list `()` a missing multidim leaf reports for a plain
 /// (non-negated) value adverb (`:v`/`:k`/`:p`/`:kv`) -- matching raku, which
 /// answers `()` rather than `Nil` for a hole under these adverbs.
-fn multidim_empty_list() -> Value {
+pub(super) fn multidim_empty_list() -> Value {
     Value::array_with_kind(
         crate::gc::Gc::new(crate::value::ArrayData::new(vec![])),
         ArrayKind::List,
@@ -37,7 +46,7 @@ fn multidim_empty_list() -> Value {
 ///   the roast-pinned answer.
 ///
 /// `:kv` does not use this -- it is always `()` for every kind of miss.
-fn multidim_missing_result(raw_value: &Value) -> Value {
+pub(super) fn multidim_missing_result(raw_value: &Value) -> Value {
     if raw_value.is_nil() {
         Value::NIL
     } else {
@@ -46,429 +55,6 @@ fn multidim_missing_result(raw_value: &Value) -> Value {
 }
 
 impl Interpreter {
-    /// Handle dynamic adverbs on multidim index: @array[$a;$b;$c]:$delete
-    /// Args: [inner_expr_result, adverb_name, adverb_value]
-    pub(super) fn builtin_multidim_adverb(
-        &mut self,
-        args: &[Value],
-    ) -> Result<Value, RuntimeError> {
-        if args.len() < 3 {
-            return Err(RuntimeError::new(
-                "__mutsu_multidim_adverb expects value, adverb_name, and adverb_value",
-            ));
-        }
-        let value = args[0].clone();
-        let _adverb_name = args[1].to_string_value();
-        let adverb_value = &args[2];
-
-        // If the adverb is False, just return the value unchanged
-        if !adverb_value.truthy() {
-            return Ok(value);
-        }
-
-        // Adverb is True — currently only "delete" is supported.
-        // When the inner expression is a MultiDimIndex result, we need to
-        // delete the element from the array. However, the value has already
-        // been evaluated, so we return it as-is for now.
-        // TODO: Implement actual delete by restructuring the parser to
-        // pass target array info.
-        Ok(value)
-    }
-
-    /// Handle subscript adverbs (:kv, :k, :v, :p, etc.) on multidim index.
-    /// Args: [target_array, adverb_name, dim0, dim1, dim2, ...]
-    pub(super) fn builtin_multidim_subscript_adverb(
-        &mut self,
-        args: &[Value],
-    ) -> Result<Value, RuntimeError> {
-        if args.len() < 3 {
-            return Err(RuntimeError::new(
-                "__mutsu_multidim_subscript_adverb expects target, adverb, and indices",
-            ));
-        }
-        let target = &args[0];
-        let adverb = args[1].to_string_value();
-        let raw_indices = &args[2..];
-        let indices = self.resolve_multidim_indices(target, raw_indices)?;
-
-        // Check if we need multi-result mode
-        if has_multi_indices(&indices) {
-            return self.multidim_subscript_adverb_multi(target, &adverb, &indices);
-        }
-
-        let (value, is_hole) = multidim_index_with_hole(target, &indices);
-        let key = make_key_tuple(&indices);
-        // Canonical hole predicate (ADR-0049 §1.6/§4 slice 5): a leaf
-        // "exists" only if it is neither a missing Hash key (`Nil`) nor an
-        // Array hole per `ArrayData::hole_at`.
-        let exists = !value.is_nil() && !is_hole;
-
-        // A missing leaf (`!exists`) reports differently depending on WHICH
-        // kind of "missing" it is, and on the adverb -- see
-        // `multidim_missing_result`'s doc comment for the full rule and its
-        // `raku`/roast evidence.
-        //
-        // The negated forms (`:!v`/`:!k`/`:!p`/`:!kv`) are a genuine Rakudo
-        // multidim quirk, also verified directly: unlike the single-dimension
-        // form (where each negated adverb keeps its own key/pair/kv shape and
-        // only suppresses the *suppression*, e.g. `:!k` on a hole still
-        // reports the key), real Rakudo's multidim `[;]` postcircumfix
-        // collapses ALL FOUR negated adverbs to plain value access --
-        // `@a[i;j]:!k` and `@a[i;j]:!p` both answer the same raw value
-        // `:!v` would, never the key or a pair, for both a filled slot and a
-        // hole.
-        match adverb.as_str() {
-            "k" => Ok(if exists {
-                key
-            } else {
-                multidim_missing_result(&value)
-            }),
-            "kv" => {
-                if exists {
-                    let v = array_to_list(value);
-                    Ok(Value::array_with_kind(
-                        crate::gc::Gc::new(crate::value::ArrayData::new(vec![key, v])),
-                        ArrayKind::List,
-                    ))
-                } else {
-                    Ok(multidim_empty_list())
-                }
-            }
-            "p" => {
-                if exists {
-                    let v = array_to_list(value);
-                    Ok(Value::value_pair(key, v))
-                } else {
-                    Ok(multidim_missing_result(&value))
-                }
-            }
-            "v" => {
-                if exists {
-                    Ok(array_to_list(value))
-                } else {
-                    Ok(multidim_missing_result(&value))
-                }
-            }
-            "not-k" | "not-kv" | "not-p" | "not-v" => Ok(array_to_list(value)),
-            _ => Ok(value),
-        }
-    }
-
-    /// Multi-result adverb handler for Whatever/list indices.
-    fn multidim_subscript_adverb_multi(
-        &mut self,
-        target: &Value,
-        adverb: &str,
-        indices: &[Value],
-    ) -> Result<Value, RuntimeError> {
-        let mut leaves = Vec::new();
-        multidim_collect_leaves(target, indices, &[], &mut leaves);
-
-        let mut out = Vec::new();
-        for (path, value, is_hole) in leaves {
-            // Canonical hole predicate (ADR-0049 §1.6/§4 slice 5): a leaf
-            // "exists" only if it is neither a missing Hash key (`Nil`) nor
-            // an Array hole per `ArrayData::hole_at`.
-            let exists = !value.is_nil() && !is_hole;
-            let key = leaf_key_tuple(path);
-            match adverb {
-                "k" => {
-                    if exists {
-                        out.push(key);
-                    }
-                }
-                "kv" => {
-                    if exists {
-                        out.push(key);
-                        out.push(array_to_list(value));
-                    }
-                }
-                "p" => {
-                    if exists {
-                        out.push(Value::value_pair(key, array_to_list(value)));
-                    }
-                }
-                "v" => {
-                    if exists {
-                        out.push(array_to_list(value));
-                    }
-                }
-                // Self-consistency with `builtin_multidim_subscript_adverb`'s
-                // single-coordinate form above (raku itself has no oracle for
-                // this Whatever/list-index combination -- it throws X::NYI --
-                // so the two mutsu code paths must at least agree with each
-                // other): a negated adverb keeps every leaf, filled or hole,
-                // reported as the raw value, never the key or a pair.
-                "not-k" | "not-kv" | "not-p" | "not-v" => {
-                    out.push(array_to_list(value));
-                }
-                _ => out.push(value),
-            }
-        }
-        Ok(Value::array(out))
-    }
-
-    /// Handle :exists with secondary adverbs on multidim index.
-    /// Args: [target_array, negated_bool, adverb_name, dim0, dim1, ...]
-    pub(super) fn builtin_multidim_exists_adverb(
-        &mut self,
-        args: &[Value],
-    ) -> Result<Value, RuntimeError> {
-        if args.len() < 4 {
-            return Err(RuntimeError::new(
-                "__mutsu_multidim_exists_adverb expects target, negated, adverb, and indices",
-            ));
-        }
-        let target = &args[0];
-        let negated = args[1].truthy();
-        let adverb = args[2].to_string_value();
-        let raw_indices = &args[3..];
-        let indices = self.resolve_multidim_indices(target, raw_indices)?;
-        // A nested single-dimension slice (`@a[(3, (30, (5,)))]:exists`) is a
-        // structure-preserving slice, NOT a multidim coordinate walk: recurse the
-        // index tree and report an existence Bool per leaf, keeping the nesting.
-        if indices.len() == 1
-            && let Some(inner) = Self::nested_index_elements(&indices[0])
-            && inner
-                .iter()
-                .any(|e| Self::nested_index_elements(e).is_some())
-            && let Some(items) = Self::positional_exists_items(target)
-        {
-            let out = Self::nested_exists_slice(&items, &inner, negated, &adverb);
-            return Ok(Value::array(out));
-        }
-        if let ValueView::Instance {
-            class_name,
-            attributes,
-            ..
-        } = target.view()
-            && class_name == "Stash"
-            && let Some(ValueView::Hash(symbols)) =
-                attributes.as_map().get("symbols").map(Value::view)
-        {
-            let stash_exists = |idx: &Value| {
-                let key = idx.to_string_value();
-                if symbols.contains_key(&key) {
-                    return true;
-                }
-                if !key.starts_with('$')
-                    && !key.starts_with('@')
-                    && !key.starts_with('%')
-                    && !key.starts_with('&')
-                {
-                    return symbols.contains_key(&format!("${key}"));
-                }
-                false
-            };
-            let stash_indices: Vec<Value> = if indices.len() > 1 {
-                indices.clone()
-            } else {
-                match indices[0].view() {
-                    ValueView::Array(items, ..) => items.to_vec(),
-                    _ => vec![indices[0].clone()],
-                }
-            };
-            let exists_vals: Vec<bool> = stash_indices.iter().map(stash_exists).collect();
-            let exists_vals: Vec<bool> = if negated {
-                exists_vals.into_iter().map(|v| !v).collect()
-            } else {
-                exists_vals
-            };
-            if stash_indices.len() > 1 {
-                return Ok(Value::array(
-                    exists_vals
-                        .into_iter()
-                        .map(Value::truth)
-                        .collect::<Vec<_>>(),
-                ));
-            }
-            return Ok(Value::truth(*exists_vals.first().unwrap_or(&false)));
-        }
-
-        // Multi-result mode for Whatever/list indices
-        if has_multi_indices(&indices) {
-            return self.multidim_exists_adverb_multi(target, negated, &adverb, &indices);
-        }
-
-        let (value, is_hole) = multidim_index_with_hole(target, &indices);
-        // Canonical hole predicate (ADR-0049 §1.6/§4 slice 5): an unassigned
-        // array cell (per `ArrayData::hole_at`, which recognizes both the
-        // untyped `Any` marker and a typed array's own element-type marker,
-        // and consults `initialized` so an explicitly-assigned `Any`/type
-        // object is NOT a hole) or a missing Hash key (`Value::NIL`) does not
-        // exist yet -- for shaped AND non-shaped (autoviv) multidim arrays
-        // alike.
-        let raw_exists = !value.is_nil() && !is_hole;
-        let exists = if negated { !raw_exists } else { raw_exists };
-        let key = make_key_tuple(&indices);
-
-        match adverb.as_str() {
-            "none" => Ok(Value::truth(exists)),
-            "kv" => {
-                if raw_exists {
-                    Ok(Value::array_with_kind(
-                        crate::gc::Gc::new(crate::value::ArrayData::new(vec![
-                            key,
-                            Value::truth(exists),
-                        ])),
-                        ArrayKind::List,
-                    ))
-                } else {
-                    Ok(Value::array_with_kind(
-                        crate::gc::Gc::new(crate::value::ArrayData::new(vec![])),
-                        ArrayKind::List,
-                    ))
-                }
-            }
-            "p" => {
-                if raw_exists {
-                    Ok(Value::value_pair(key, Value::truth(exists)))
-                } else {
-                    Ok(Value::NIL)
-                }
-            }
-            "k" => {
-                if raw_exists {
-                    Ok(key)
-                } else {
-                    Ok(Value::NIL)
-                }
-            }
-            "v" => Ok(Value::truth(exists)),
-            _ => Ok(Value::truth(exists)),
-        }
-    }
-
-    /// The positional element view (with unfilled-slot info folded into the
-    /// values) for a nested `:exists` slice, or `None` for a non-positional
-    /// target. A real array keeps its hole marks (`Package("Any")` in an
-    /// uninitialized slot); a Range/List is fully filled.
-    pub(crate) fn positional_exists_items(target: &Value) -> Option<Vec<Value>> {
-        match target.view() {
-            ValueView::Array(items, ..) => Some(items.to_vec()),
-            _ if target.is_range()
-                || matches!(
-                    target.view(),
-                    ValueView::Seq(_)
-                        | ValueView::HyperSeq(_)
-                        | ValueView::RaceSeq(_)
-                        | ValueView::LazyList(_)
-                ) =>
-            {
-                Some(crate::runtime::utils::value_to_list(target))
-            }
-            _ => None,
-        }
-    }
-
-    /// One level of a nested `:exists`/`:!exists` slice: a sub-list index recurses
-    /// into ONE nested list element; a scalar index reports whether that slot
-    /// exists (negated by `:!exists`), formatted by the companion adverb
-    /// (`none`/`k`/`v`/`kv`/`p`).
-    pub(crate) fn nested_exists_slice(
-        items: &[Value],
-        indices: &[Value],
-        negated: bool,
-        adverb: &str,
-    ) -> Vec<Value> {
-        let mut out = Vec::new();
-        for idx in indices {
-            if let Some(sub) = Self::nested_index_elements(idx) {
-                out.push(Value::array(Self::nested_exists_slice(
-                    items, &sub, negated, adverb,
-                )));
-                continue;
-            }
-            let i = match idx.view() {
-                ValueView::Int(i) => i,
-                ValueView::Num(f) => f as i64,
-                _ => idx.to_string_value().parse::<i64>().unwrap_or(-1),
-            };
-            let raw_exists = i >= 0
-                && (i as usize) < items.len()
-                && !matches!(items[i as usize].view(), ValueView::Package(name) if name == "Any");
-            let exists = if negated { !raw_exists } else { raw_exists };
-            let key = Value::int(i);
-            match adverb {
-                // `:k` / `:kv` / `:p` keep only actually-existing keys.
-                "k" => {
-                    if raw_exists {
-                        out.push(key);
-                    }
-                }
-                "kv" => {
-                    if raw_exists {
-                        out.push(key);
-                        out.push(Value::truth(exists));
-                    }
-                }
-                "p" => {
-                    if raw_exists {
-                        out.push(Value::value_pair(key, Value::truth(exists)));
-                    }
-                }
-                // `:!kv` / `:!p` keep every attempted key.
-                "not-kv" => {
-                    out.push(key);
-                    out.push(Value::truth(exists));
-                }
-                "not-p" => {
-                    out.push(Value::value_pair(key, Value::truth(exists)));
-                }
-                // `none` and `v` report a Bool for every index.
-                _ => out.push(Value::truth(exists)),
-            }
-        }
-        out
-    }
-
-    /// Multi-result :exists adverb handler for Whatever/list indices.
-    fn multidim_exists_adverb_multi(
-        &mut self,
-        target: &Value,
-        negated: bool,
-        adverb: &str,
-        indices: &[Value],
-    ) -> Result<Value, RuntimeError> {
-        let mut leaves = Vec::new();
-        multidim_collect_leaves(target, indices, &[], &mut leaves);
-
-        let mut out = Vec::new();
-        for (path, value, is_hole) in leaves {
-            // Canonical hole predicate (ADR-0049 §1.6/§4 slice 5): a
-            // deleted/uninitialized array slot (per `ArrayData::hole_at`,
-            // which also recognizes a typed array's own element-type gap
-            // marker and consults `initialized` so an explicitly-assigned
-            // `Any`/type-object value is NOT treated as a hole) or a missing
-            // Hash key (`Value::NIL`) reports False, not True.
-            let raw_exists = !value.is_nil() && !is_hole;
-            let exists = if negated { !raw_exists } else { raw_exists };
-            let key = leaf_key_tuple(path);
-            match adverb {
-                "none" => out.push(Value::truth(exists)),
-                "kv" => {
-                    if raw_exists {
-                        out.push(key);
-                        out.push(Value::truth(exists));
-                    }
-                }
-                "p" => {
-                    if raw_exists {
-                        out.push(Value::value_pair(key, Value::truth(exists)));
-                    }
-                }
-                "k" => {
-                    if raw_exists {
-                        out.push(key);
-                    }
-                }
-                "v" => out.push(Value::truth(exists)),
-                _ => out.push(Value::truth(exists)),
-            }
-        }
-        Ok(Value::array(out))
-    }
-
     /// Handle :delete on multidim index.
     /// Args: [var_name_string, dim0, dim1, ...]
     pub(super) fn builtin_multidim_delete(
@@ -595,7 +181,7 @@ impl Interpreter {
 
     /// Resolve WhateverCode indices: if an index is a Sub (WhateverCode),
     /// call it with the current dimension's array length.
-    fn resolve_multidim_indices(
+    pub(super) fn resolve_multidim_indices(
         &mut self,
         target: &Value,
         indices: &[Value],
@@ -683,296 +269,5 @@ impl Interpreter {
             }
         }
         Ok(resolved)
-    }
-
-    /// Handle dynamic adverb (:$delete) on multidim index.
-    /// Args: [var_name_string, adverb_name, adverb_value, dim0, dim1, ...]
-    pub(super) fn builtin_multidim_dynamic_adverb(
-        &mut self,
-        args: &mut [Value],
-    ) -> Result<Value, RuntimeError> {
-        if args.len() < 4 {
-            return Err(RuntimeError::new(
-                "__mutsu_multidim_dynamic_adverb expects var_name, name, value, and indices",
-            ));
-        }
-        let var_name = args[0].to_string_value();
-        let adverb_value = args[2].truthy();
-        let raw_indices = args[3..].to_vec();
-
-        let target = self.env.get(&var_name).cloned().unwrap_or(Value::NIL);
-        let indices = self.resolve_multidim_indices(&target, &raw_indices)?;
-
-        if adverb_value {
-            // Multi-result mode for Whatever/list indices
-            if has_multi_indices(&indices) {
-                let mut leaves = Vec::new();
-                multidim_collect_leaves(&target, &indices, &[], &mut leaves);
-                if let Some(t) = self.env.get_mut(&var_name) {
-                    multidim_delete(t, &indices);
-                    self.writeback_multidim_var_to_local(&var_name);
-                }
-                let values: Vec<Value> = leaves
-                    .into_iter()
-                    .map(|(_, v, _)| array_to_list(v))
-                    .collect();
-                return Ok(Value::array(values));
-            }
-            // A non-existent (out-of-range) element deletes to `Nil`, not the
-            // `Any` hole-value that `multidim_delete` returns for a missing slot.
-            if multidim_index(&target, &indices).is_nil() {
-                return Ok(Value::NIL);
-            }
-            let Some(target) = self.env.get_mut(&var_name) else {
-                return Ok(Value::NIL);
-            };
-            let result = multidim_delete(target, &indices);
-            self.writeback_multidim_var_to_local(&var_name);
-            // The deleted element is returned decontainerized — an Array leaf
-            // (`[314]`) comes back as a List (`(314,)`), matching raku's
-            // multi-dim `:delete` (the test's `$resnona` is `$result.List`).
-            Ok(array_to_list(result))
-        } else {
-            Ok(multidim_index(&target, &indices))
-        }
-    }
-
-    /// Handle :kv/:k/:v/:p with dynamic :$delete on multidim index.
-    /// Args: [var_name_str, adverb_name, delete_flag, dim0, dim1, ...]
-    pub(super) fn builtin_multidim_subscript_adverb_dyn(
-        &mut self,
-        args: &mut [Value],
-    ) -> Result<Value, RuntimeError> {
-        if args.len() < 4 {
-            return Err(RuntimeError::new(
-                "__mutsu_multidim_subscript_adverb_dyn expects var_name, adverb, delete, and indices",
-            ));
-        }
-        let var_name = args[0].to_string_value();
-        let adverb = args[1].to_string_value();
-        let do_delete = args[2].truthy();
-        let raw_indices = args[3..].to_vec();
-
-        let target = self.env.get(&var_name).cloned().unwrap_or(Value::NIL);
-        let indices = self.resolve_multidim_indices(&target, &raw_indices)?;
-
-        // Multi-result mode for Whatever/list indices
-        if has_multi_indices(&indices) {
-            // Collect leaves before potentially deleting
-            let mut leaves = Vec::new();
-            multidim_collect_leaves(&target, &indices, &[], &mut leaves);
-            if do_delete && let Some(t) = self.env.get_mut(&var_name) {
-                multidim_delete(t, &indices);
-                self.writeback_multidim_var_to_local(&var_name);
-            }
-            let mut out = Vec::new();
-            for (path, value, is_hole) in leaves {
-                // Canonical hole predicate (ADR-0049 §1.6/§4 slice 5) -- see
-                // `multidim_exists_adverb_multi` above.
-                let exists = !value.is_nil() && !is_hole;
-                let key = leaf_key_tuple(path);
-                match adverb.as_str() {
-                    "k" => {
-                        if exists {
-                            out.push(key);
-                        }
-                    }
-                    "kv" => {
-                        if exists {
-                            out.push(key);
-                            out.push(array_to_list(value));
-                        }
-                    }
-                    "p" => {
-                        if exists {
-                            out.push(Value::value_pair(key, array_to_list(value)));
-                        }
-                    }
-                    "v" => {
-                        if exists {
-                            out.push(array_to_list(value));
-                        }
-                    }
-                    // Self-consistency with the plain (non-dyn) handlers above:
-                    // a negated adverb keeps every leaf, filled or hole,
-                    // reported as the raw value.
-                    "not-k" | "not-kv" | "not-p" | "not-v" => {
-                        out.push(array_to_list(value));
-                    }
-                    _ => out.push(value),
-                }
-            }
-            return Ok(Value::array(out));
-        }
-
-        // Determine existence from a pre-delete read: a non-existent element
-        // reads as `Nil` or an `ArrayData::hole_at` hole (ADR-0049 §1.6/§4
-        // slice 5), whereas `multidim_delete` returns the `Any` hole-value
-        // for an out-of-range slot, which would wrongly look "present".
-        let (read_value, read_is_hole) = multidim_index_with_hole(&target, &indices);
-        let exists = !read_value.is_nil() && !read_is_hole;
-        let value = if do_delete {
-            if exists && let Some(target) = self.env.get_mut(&var_name) {
-                let r = multidim_delete(target, &indices);
-                self.writeback_multidim_var_to_local(&var_name);
-                r
-            } else {
-                read_value
-            }
-        } else {
-            read_value
-        };
-
-        let key = make_key_tuple(&indices);
-
-        // Same shape/hole rules as `builtin_multidim_subscript_adverb` above
-        // (see `multidim_missing_result`'s doc comment): a missing leaf
-        // reports `()` for an in-bounds Array hole, `Nil` for everything
-        // else that fails to resolve, `:kv` is always `()`, and every
-        // negated adverb collapses to plain (raw) value access.
-        match adverb.as_str() {
-            "k" => Ok(if exists {
-                key
-            } else {
-                multidim_missing_result(&value)
-            }),
-            "kv" => {
-                if exists {
-                    let v = array_to_list(value);
-                    Ok(Value::array_with_kind(
-                        crate::gc::Gc::new(crate::value::ArrayData::new(vec![key, v])),
-                        ArrayKind::List,
-                    ))
-                } else {
-                    Ok(multidim_empty_list())
-                }
-            }
-            "p" => {
-                if exists {
-                    let v = array_to_list(value);
-                    Ok(Value::value_pair(key, v))
-                } else {
-                    Ok(multidim_missing_result(&value))
-                }
-            }
-            "v" => {
-                if exists {
-                    Ok(array_to_list(value))
-                } else {
-                    Ok(multidim_missing_result(&value))
-                }
-            }
-            "not-k" | "not-kv" | "not-p" | "not-v" => Ok(array_to_list(value)),
-            _ => Ok(value),
-        }
-    }
-
-    /// Handle :exists:kv/:exists:p with dynamic :$delete on multidim index.
-    /// Args: [var_name_str, negated_bool, delete_flag, adverb_name, dim0, dim1, ...]
-    pub(super) fn builtin_multidim_exists_adverb_dyn(
-        &mut self,
-        args: &mut [Value],
-    ) -> Result<Value, RuntimeError> {
-        if args.len() < 5 {
-            return Err(RuntimeError::new(
-                "__mutsu_multidim_exists_adverb_dyn requires 5+ args",
-            ));
-        }
-        let var_name = args[0].to_string_value();
-        let negated = args[1].truthy();
-        let do_delete = args[2].truthy();
-        let adverb = args[3].to_string_value();
-        let raw_indices = args[4..].to_vec();
-
-        // First get value (need to read before potentially deleting)
-        let target_val = self.env.get(&var_name).cloned().unwrap_or(Value::NIL);
-        let indices = self.resolve_multidim_indices(&target_val, &raw_indices)?;
-
-        // Multi-result mode for Whatever/list indices
-        if has_multi_indices(&indices) {
-            let mut leaves = Vec::new();
-            multidim_collect_leaves(&target_val, &indices, &[], &mut leaves);
-            if do_delete && let Some(t) = self.env.get_mut(&var_name) {
-                multidim_delete(t, &indices);
-                self.writeback_multidim_var_to_local(&var_name);
-            }
-            let mut out = Vec::new();
-            for (path, value, is_hole) in leaves {
-                // Canonical hole predicate (ADR-0049 §1.6/§4 slice 5) -- see
-                // `multidim_exists_adverb_multi` above.
-                let raw_exists = !value.is_nil() && !is_hole;
-                let exists = if negated { !raw_exists } else { raw_exists };
-                let key = leaf_key_tuple(path);
-                match adverb.as_str() {
-                    "none" => out.push(Value::truth(exists)),
-                    "kv" => {
-                        if raw_exists {
-                            out.push(key);
-                            out.push(Value::truth(exists));
-                        }
-                    }
-                    "p" => {
-                        if raw_exists {
-                            out.push(Value::value_pair(key, Value::truth(exists)));
-                        }
-                    }
-                    "k" => {
-                        if raw_exists {
-                            out.push(key);
-                        }
-                    }
-                    _ => out.push(Value::truth(exists)),
-                }
-            }
-            return Ok(Value::array(out));
-        }
-
-        let (value, is_hole) = multidim_index_with_hole(&target_val, &indices);
-        // Then delete if requested
-        if do_delete && let Some(target) = self.env.get_mut(&var_name) {
-            multidim_delete(target, &indices);
-            self.writeback_multidim_var_to_local(&var_name);
-        }
-
-        // Canonical hole predicate (ADR-0049 §1.6/§4 slice 5) -- see
-        // `multidim_exists_adverb_multi` above.
-        let raw_exists = !value.is_nil() && !is_hole;
-        let exists = if negated { !raw_exists } else { raw_exists };
-        let key = make_key_tuple(&indices);
-
-        match adverb.as_str() {
-            "none" => Ok(Value::truth(exists)),
-            "kv" => {
-                if raw_exists {
-                    Ok(Value::array_with_kind(
-                        crate::gc::Gc::new(crate::value::ArrayData::new(vec![
-                            key,
-                            Value::truth(exists),
-                        ])),
-                        ArrayKind::List,
-                    ))
-                } else {
-                    Ok(Value::array_with_kind(
-                        crate::gc::Gc::new(crate::value::ArrayData::new(vec![])),
-                        ArrayKind::List,
-                    ))
-                }
-            }
-            "p" => {
-                if raw_exists {
-                    Ok(Value::value_pair(key, Value::truth(exists)))
-                } else {
-                    Ok(Value::NIL)
-                }
-            }
-            "k" => {
-                if raw_exists {
-                    Ok(key)
-                } else {
-                    Ok(Value::NIL)
-                }
-            }
-            _ => Ok(Value::truth(exists)),
-        }
     }
 }
