+++ src/runtime/mod.rs
@@ -393,8 +393,10 @@ mod builtins_io_stream;
 mod builtins_lvalue;
 mod builtins_multidim;
 mod builtins_multidim_assign;
+mod builtins_multidim_exists_adverb;
 mod builtins_multidim_ops;
 mod builtins_multidim_subscript;
+mod builtins_multidim_subscript_adverb;
 pub(crate) use builtins_multidim_subscript::PositionalMissing;
 mod builtins_operators_coerce;
 mod builtins_operators_fallback;
